# Optimized all `Rect/FRect` methods via `pgRect_FromObject`