@itzpr3d4t0r 2024-06-09
I must report that this chage has drawbacks, specifically with Rect/Frect subclasses. You can easily see it from this graph, so even tho we see significant improvements in basically all cases, rect subclasses get significantly slower. What i think is that's while it's cool to be able to subclass Rect/Frect we shouldn't prioritize its performance against most prevalent usecases such as direct Rect/Frect usage or tuples passed as rects.
![image](https://github.com/pygame-community/pygame-ce/assets/103119829/56ab37fa-6a90-4a39-8438-4cf915862901)